+++ server/responses/task_queued.go
@@ -2,7 +2,7 @@ package responses
 
 // API generate simply returns a UUID to track the request to our compute while its in flight
 type TaskQueuedResponse struct {
-	ID               string  `json:"id"`
-	UIId             string  `json:"ui_id,omitempty"`
-	RemainingCredits float32 `json:"total_remaining_credits"`
+	ID               string `json:"id"`
+	UIId             string `json:"ui_id,omitempty"`
+	RemainingCredits int    `json:"total_remaining_credits"`
 }
